test of an idea i had
