+++ include/mqt-core/dd/DensityDDContainer.hpp
@@ -1,0 +1,23 @@
+#pragma once
+
+#include "dd/DesicionDiagramContainer.hpp"
+
+namespace dd {
+
+class DensityDDContainer : public DDContainer<dNode> {
+public:
+  using DDContainer<dNode>::DDContainer;
+
+    /**
+   * @brief Construct the all-zero density operator
+            \f$|0...0\rangle\langle0...0|\f$
+   * @param n The number of qubits
+   * @return A decision diagram for the all-zero density operator
+   */
+  dEdge makeZeroDensityOperator(std::size_t n);
+
+  char measureOneCollapsing(dEdge& e, Qubit index, std::mt19937_64& mt);
+
+};
+
+} // namespace dd
