+++ test/algorithms/eval_dynamic_circuits.cpp
@@ -126,7 +126,7 @@ TEST_P(DynamicCircuitEvalExactQPE, UnitaryTransformation) {
   iqpe.reorderOperations();
   const auto finishedTransformation = std::chrono::steady_clock::now();
 
-  dd::MatrixDD e = dd::Package::makeIdent();
+  dd::MatrixDD e = dd::MatrixDDContainer::makeIdent();
   dd->incRef(e);
 
   auto leftIt = qpe.begin();
@@ -295,7 +295,7 @@ TEST_P(DynamicCircuitEvalInexactQPE, UnitaryTransformation) {
   iqpe.reorderOperations();
   const auto finishedTransformation = std::chrono::steady_clock::now();
 
-  dd::MatrixDD e = dd::Package::makeIdent();
+  dd::MatrixDD e = dd::MatrixDDContainer::makeIdent();
   dd->incRef(e);
 
   auto leftIt = qpe.begin();
@@ -408,7 +408,7 @@ TEST_P(DynamicCircuitEvalBV, UnitaryTransformation) {
   dbv.reorderOperations();
   const auto finishedTransformation = std::chrono::steady_clock::now();
 
-  dd::MatrixDD e = dd::Package::makeIdent();
+  dd::MatrixDD e = dd::MatrixDDContainer::makeIdent();
   dd->incRef(e);
 
   auto leftIt = bv.begin();
@@ -517,7 +517,7 @@ TEST_P(DynamicCircuitEvalQFT, UnitaryTransformation) {
   dqft.reorderOperations();
   const auto finishedTransformation = std::chrono::steady_clock::now();
 
-  dd::MatrixDD e = dd::Package::makeIdent();
+  dd::MatrixDD e = dd::MatrixDDContainer::makeIdent();
   dd->incRef(e);
 
   auto leftIt = qft.begin();
