+++ test/algorithms/test_bernsteinvazirani.cpp
@@ -122,7 +122,7 @@ TEST_P(BernsteinVazirani, DynamicEquivalenceSimulation) {
   qc::CircuitOptimizer::removeFinalMeasurements(bv);
 
   // simulate circuit
-  auto e = dd::simulate(bv, dd->makeZeroState(bv.getNqubits()), *dd);
+  auto e = dd::simulate(bv, dd->vectors().makeZeroState(bv.getNqubits()), *dd);
 
   // create dynamic BV circuit
   auto dbv = qc::createIterativeBernsteinVazirani(s);
@@ -137,9 +137,9 @@ TEST_P(BernsteinVazirani, DynamicEquivalenceSimulation) {
   qc::CircuitOptimizer::removeFinalMeasurements(dbv);
 
   // simulate circuit
-  auto f = dd::simulate(dbv, dd->makeZeroState(dbv.getNqubits()), *dd);
+  auto f = dd::simulate(dbv, dd->vectors().makeZeroState(dbv.getNqubits()), *dd);
 
   // calculate fidelity between both results
-  auto fidelity = dd->fidelity(e, f);
+  auto fidelity = dd->vectors().fidelity(e, f);
   EXPECT_NEAR(fidelity, 1.0, 1e-4);
 }
