+++ test/algorithms/test_entanglement.cpp
@@ -45,23 +45,23 @@ TEST_P(Entanglement, FunctionTest) {
   const auto qc = qc::createGHZState(nq);
   const auto e = dd::buildFunctionality(qc, *dd);
   ASSERT_EQ(qc.getNops(), nq);
-  const auto r = dd->multiply(e, dd->makeZeroState(nq));
+  const auto r = dd->multiply(e, dd->vectors().makeZeroState(nq));
   ASSERT_EQ(r.getValueByPath(nq, std::string(nq, '0')), dd::SQRT2_2);
   ASSERT_EQ(r.getValueByPath(nq, std::string(nq, '1')), dd::SQRT2_2);
 }
 
 TEST_P(Entanglement, GHZRoutineFunctionTest) {
   const auto qc = qc::createGHZState(nq);
-  const auto e = dd::simulate(qc, dd->makeZeroState(nq), *dd);
-  const auto f = dd->makeGHZState(nq);
+  const auto e = dd::simulate(qc, dd->vectors().makeZeroState(nq), *dd);
+  const auto f = dd->vectors().makeGHZState(nq);
   EXPECT_EQ(e, f);
 }
 
 TEST(Entanglement, GHZStateEdgeCasesTest) {
   auto dd = std::make_unique<dd::Package>(3);
 
-  EXPECT_EQ(dd->makeGHZState(0),
-            dd->makeBasisState(0, {dd::BasisStates::zero}));
-  EXPECT_EQ(dd->makeGHZState(0), dd->makeBasisState(0, {dd::BasisStates::one}));
-  ASSERT_THROW(dd->makeGHZState(6), std::runtime_error);
+  EXPECT_EQ(dd->vectors().makeGHZState(0),
+            dd->vectors().makeBasisState(0, {dd::BasisStates::zero}));
+  EXPECT_EQ(dd->vectors().makeGHZState(0), dd->vectors().makeBasisState(0, {dd::BasisStates::one}));
+  ASSERT_THROW(dd->vectors().makeGHZState(6), std::runtime_error);
 }
