+++ test/algorithms/test_statepreparation.cpp
@@ -65,7 +65,7 @@ TEST_P(StatePreparation, StatePreparationCircuitSimulation) {
   auto dd = std::make_unique<dd::Package>(qc.getNqubits());
   dd::VectorDD e{};
   ASSERT_NO_THROW(
-      { e = dd::simulate(qc, dd->makeZeroState(qc.getNqubits()), *dd); });
+      { e = dd::simulate(qc, dd->vectors().makeZeroState(qc.getNqubits()), *dd); });
   auto result = e.getVector();
   for (size_t i = 0; i < expectedAmplitudes.size(); ++i) {
     ASSERT_NEAR(expectedAmplitudes[i].real(), result[i].real(), EPS);
