+++ test/dd/test_dd_noise_functionality.cpp
@@ -79,7 +79,7 @@ TEST_F(DDNoiseFunctionalityTest, DetSimulateAdder4TrackAPD) {
   auto dd = std::make_unique<dd::Package>(
       qc.getNqubits(), dd::DENSITY_MATRIX_SIMULATOR_DD_PACKAGE_CONFIG);
 
-  auto rootEdge = dd->makeZeroDensityOperator(qc.getNqubits());
+  auto rootEdge = dd->densities().makeZeroDensityOperator(qc.getNqubits());
 
   const auto* const noiseEffects = "APDI";
 
@@ -111,7 +111,7 @@ TEST_F(DDNoiseFunctionalityTest, DetSimulateAdder4TrackD) {
   auto dd = std::make_unique<dd::Package>(
       qc.getNqubits(), dd::DENSITY_MATRIX_SIMULATOR_DD_PACKAGE_CONFIG);
 
-  auto rootEdge = dd->makeZeroDensityOperator(qc.getNqubits());
+  auto rootEdge = dd->densities().makeZeroDensityOperator(qc.getNqubits());
 
   const auto* const noiseEffects = "D";
 
@@ -143,7 +143,7 @@ TEST_F(DDNoiseFunctionalityTest, testingMeasure) {
   auto dd = std::make_unique<dd::Package>(
       qcOp.getNqubits(), dd::DENSITY_MATRIX_SIMULATOR_DD_PACKAGE_CONFIG);
 
-  auto rootEdge = dd->makeZeroDensityOperator(qcOp.getNqubits());
+  auto rootEdge = dd->densities().makeZeroDensityOperator(qcOp.getNqubits());
 
   auto deterministicNoiseFunctionality = dd::DeterministicNoiseFunctionality(
       *dd, qcOp.getNqubits(), 0.01, 0.02, 0.02, 0.04, {});
@@ -209,7 +209,7 @@ TEST_F(DDNoiseFunctionalityTest, StochSimulateAdder4TrackAPD) {
       *dd, qc.getNqubits(), 0.01, 0.02, 2., noiseEffects);
 
   for (size_t i = 0U; i < stochRuns; i++) {
-    auto rootEdge = dd->makeZeroState(qc.getNqubits());
+    auto rootEdge = dd->vectors().makeZeroState(qc.getNqubits());
     dd->incRef(rootEdge);
 
     for (auto const& op : qc) {
@@ -262,7 +262,7 @@ TEST_F(DDNoiseFunctionalityTest, StochSimulateAdder4IdentityError) {
       *dd, qc.getNqubits(), 0.01, 0.02, 2., noiseEffects);
 
   for (size_t i = 0U; i < stochRuns; i++) {
-    auto rootEdge = dd->makeZeroState(qc.getNqubits());
+    auto rootEdge = dd->vectors().makeZeroState(qc.getNqubits());
     dd->incRef(rootEdge);
 
     for (auto const& op : qc) {
